requirements.txt
mysql-connector-python==8.0.33
confluent-kafka==2.3.0
watchdog==3.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.4.2 